requirements.txt
# ── Web framework ─────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1

# ── Image processing ──────────────────────────────────────────────
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
scikit-image==0.25.2

# ── Deep learning ─────────────────────────────────────────────────
tensorflow==2.16.1

# ── Auth & Database ───────────────────────────────────────────────
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
email-validator==2.1.1

# ── PDF generation ────────────────────────────────────────────────
reportlab==4.0.8
